I understood it at the Third Germany in the cards. The identity of devil in clinton. This is for us?
It is America that helped Chiang Kai-shek and made China a communist zone. Turning back to that American history, you will find out that Russia will be the G20. It is a braking for President Trump to  Syrian war.It is a form avoided against Europe. The peace treaty with Russia in Japan itself wrong.
I do not think that we can live our safe living in Japan if  can go through the Pacific. It is annoying to us that the system is up there, right? Safety is protected only by our living. I do not think to be good  the alliance of the three countries is also for Japan again.